% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.EsoralPrint.R
\name{predict.EsoralPrint}
\alias{predict.EsoralPrint}
\title{Predict Method for EsoralPrint}
\usage{
\method{predict}{EsoralPrint}(object, newdata, type = c("prob", "class"), cutoff = 0.5, ...)
}
\arguments{
\item{object}{EsoralPrint model}

\item{newdata}{Numeric matrix with genes as rows, samples as columns}

\item{type}{"prob" or "class"}

\item{cutoff}{numeric threshold for classification}

\item{...}{Optional additional arguments (currently unused)}
}
\value{
Data.frame with LinearScore, MalignancyProbability, and optionally Class.
}
\description{
Predict Method for EsoralPrint
}
